refactor: simplify build and packaging of progress
