+++ packages/components/progress/bs-progress.js
@@ -1,6 +1,6 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsProgressCss } from './bs-progress.css.js';
+import { BsProgressCss } from './css/bs-progress.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 export class BsProgress extends LitElement {
