+++ packages/components/progress/css/index.js
@@ -1,4 +1,2 @@
 export { BsProgressBarCss } from './bs-progress-bar.css.js';
-export { BsProgressBar } from './bs-progress-bar.js';
 export { BsProgressCss } from './bs-progress.css.js';
-export { BsProgress } from './bs-progress.js';
