+++ packages/components/progress/index.js
@@ -1,0 +1,3 @@
+export * from './css'
+export { BsProgressBar } from './bs-progress-bar.js';
+export { BsProgress } from './bs-progress.js';
